Fix Sell, Product validate, Datatype

+++ PRO1041-PRO1041.03_KieuMai-GROUP3/src/main/java/haladesign/Utitlity/ValidateNumber.java
@@ -11,7 +11,7 @@ public class ValidateNumber {
 
     public Boolean isNumber(String number) {
         try {
-            Integer.valueOf(number);
+            Long.valueOf(number);
             return true;
         } catch (NumberFormatException e) {
             return false;
